Move all functions related to clustering to new clustering file
