+++ MuoVErsi/sources/GTFS/__init__.py
@@ -1,1 +1,2 @@
+from .clustering import *
 from .source import *
